fix: theme and bill's history, add user credit rate, fix p2p and me2me transactions

See merge request HappyPiece/twomenonebank!60

+++ client_Customer/app/src/main/java/com/example/customerclient/domain/usecases/bill/WithdrawUseCase.kt
@@ -5,6 +5,6 @@ import com.example.customerclient.domain.repositories.TransactionRepository
 class WithdrawUseCase(
     private val transactionRepository: TransactionRepository
 ) {
-    suspend operator fun invoke(billId: String, amount: Int) =
+    suspend operator fun invoke(billId: String, amount: Double) =
         transactionRepository.withdraw(billId, amount)
 }
